Add clan member manager to cache clanmember ranks, use in clanchat plugin
